requirements.txt
# TTS Evaluation Framework Dependencies

# API Clients
requests>=2.31.0
cartesia>=1.0.0
elevenlabs>=1.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Audio Processing
pydub>=0.25.1

# CLI Interface
click>=8.1.7
rich>=13.7.0
questionary>=2.0.1

# Data Handling
pandas>=2.1.4
numpy>=1.26.2

# Evaluation Interface
flask>=3.0.0
gradio>=4.12.0

# Utilities
tqdm>=4.66.1
